==> zgea/10b_ema_buy_and_sell_strategie_bonds.ipynb <==
# EMA Buy and Sell Strategy (Bonds)

import pandas as pd
from pathlib import Path

from utils.plots import compare_portfolios
from utils.math import normalize_df
from utils.portfolio import Portfolio, MAPortfolio

clean_data_path = Path("clean_data")
cache_path = Path("cached_data")

input_path = clean_data_path / "etfs.xlsx"
etfs = pd.read_excel(input_path, index_col=0)
etfs.index = pd.to_datetime(etfs.index)
etfs['cash'] = 100.0
etfs

p_sp500 = Portfolio(
    {
        '1x_sp500_eu': 100.0,
    },
    start_value = 1000,
).backtest(etfs)

p_ltt = Portfolio(
    {
        '1x_ltt_eu': 100.0,
    },
    start_value = 1000,
).backtest(etfs)

p_3x_ltt = Portfolio(
    {
        '3x_ltt_us': 100.0,
    },
    start_value = 1000,
).backtest(etfs)

p_3x_itt = Portfolio(
    {
        '3x_itt_eu': 100.0,
    },
    start_value = 1000,
).backtest(etfs)

p_gold = Portfolio(
    {
        '1x_gold_eu': 100.0,
    },
    start_value = 1000,
).backtest(etfs)


start_date = '1943'
end_date = '1986'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 1x LTT")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "1x_ltt_eu": dict(dist=100, ma=ma, ma_asset="1x_ltt_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['LTT'] = p_ltt.loc[start_date:end_date]
short_names.append('LTT')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

mas=[60, 70, 90, 80, 100, 110, 120, 220, 190, 200, 180, 210]
mas=[60, 70, 90, 80, 100, 110, 120]
sum(mas)/len(mas)

start_date = '1986'
end_date = '2021'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 1x LTT")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "1x_ltt_eu": dict(dist=100, ma=ma, ma_asset="1x_ltt_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['LTT'] = p_ltt.loc[start_date:end_date]
short_names.append('LTT')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

start_date = '1943'
end_date = '1986'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 3x LTT")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "3x_ltt_us": dict(dist=100, ma=ma, ma_asset="1x_ltt_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['LTT'] = p_ltt.loc[start_date:end_date]
short_names.append('LTT')
portfolios['3x LTT'] = p_3x_ltt.loc[start_date:end_date]
short_names.append('3x LTT')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

mas=[60, 70, 80, 90]
sum(mas)/len(mas)

start_date = '1986'
end_date = '2021'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 3x LTT")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "3x_ltt_us": dict(dist=100, ma=ma, ma_asset="1x_ltt_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['LTT'] = p_ltt.loc[start_date:end_date]
short_names.append('LTT')
portfolios['3x LTT'] = p_3x_ltt.loc[start_date:end_date]
short_names.append('3x LTT')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

start_date = '1943'
end_date = '1986'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 3x ITT")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "3x_itt_eu": dict(dist=100, ma=ma, ma_asset="1x_itt_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['LTT'] = p_ltt.loc[start_date:end_date]
short_names.append('LTT')
portfolios['3x LTT'] = p_3x_ltt.loc[start_date:end_date]
short_names.append('3x LTT')
portfolios['3x ITT'] = p_3x_itt.loc[start_date:end_date]
short_names.append('3x ITT')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

mas=[40, 50, 60, 70, 80, 90, 100]
sum(mas)/len(mas)

start_date = '1986'
end_date = '2021'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 3x ITT")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "3x_itt_eu": dict(dist=100, ma=ma, ma_asset="1x_itt_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['LTT'] = p_ltt.loc[start_date:end_date]
short_names.append('LTT')
portfolios['3x LTT'] = p_3x_ltt.loc[start_date:end_date]
short_names.append('3x LTT')
portfolios['3x ITT'] = p_3x_itt.loc[start_date:end_date]
short_names.append('3x ITT')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

# EMA Buy and Sell Strategy (Gold)

start_date = '1943'
end_date = '1986'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 1x Gold")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "1x_gold_eu": dict(dist=100, ma=ma, ma_asset="1x_gold_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['Gold'] = p_gold.loc[start_date:end_date]
short_names.append('Gold')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
) 

mas = [420, 390, 450, 440, 430, 470, 410, 490, 480, 330, 350, 340, 320, 310]
sum(mas)/len(mas)

start_date = '1986'
end_date = '2021'
mas = list(range(40,500,10))

short_names = []
portfolios = {}
for ma in mas:
    name = f"MA{ma}"
    print(f"Calculate: {name} for 1x Gold")
    short_names.append(name)
    portfolios[name] = MAPortfolio(
        {
            "1x_gold_eu": dict(dist=100, ma=ma, ma_asset="1x_gold_eu"),
        },
        start_value = 1000,
    ).backtest(etfs).loc[start_date:end_date]
    

portfolios['S&P500'] = p_sp500.loc[start_date:end_date]
short_names.append('S&P500')
portfolios['Gold'] = p_gold.loc[start_date:end_date]
short_names.append('Gold')

for v in portfolios.values():
    v = normalize_df(v, start_value = 1000)

compare_portfolios(
    portfolios,
    short_names = short_names,
    details=True,
)
