==> zgea/13_simulate_monthly_rate.ipynb <==
# Simulate Monthly Rate

import pandas as pd
from pathlib import Path
from dateutil.relativedelta import relativedelta

from utils.math import perform_monte_carlo_simulation
from utils.portfolio import GermanTaxModel, MAPortfolio, Portfolio

# load raw and clean data
clean_data_path = Path("clean_data")
cache_path = Path("cached_data")

# get preprocessed asset data and convert to have useable index
input_path = clean_data_path / "etfs.xlsx"
etfs = pd.read_excel(input_path, index_col=0)
etfs.index = pd.to_datetime(etfs.index)
etfs['cash'] = 100.0
etfs.head()

p_cash = MAPortfolio(
    {
        'cash': dict(dist=100),
    },
    start_value = 10000,
).backtest(etfs)

p_50 = MAPortfolio(
    {
        '2x_sp500_eu': dict(dist=50),
        '1x_ltt_eu': dict(dist=50),
    },
    start_value = 10000,
    rebalancing = relativedelta(months=3),
    rebalancing_offset = relativedelta(days=-8), 
    spread = 0.002,
    tax_model=GermanTaxModel(),
).backtest(etfs)

p_80 = MAPortfolio(
    {
        '2x_sp500_eu': dict(dist=80),
        '1x_ltt_eu': dict(dist=20),
    },
    start_value = 10000,
    rebalancing = relativedelta(months=3),
    rebalancing_offset = relativedelta(days=-8), 
    spread = 0.002,
    tax_model=GermanTaxModel(),
).backtest(etfs)

p_65_3x = MAPortfolio(
    {
        '3x_sp500_eu': dict(dist=65),
        '3x_itt_eu': dict(dist=35),
    },
    start_value = 10000,
    rebalancing = relativedelta(months=3),
    rebalancing_offset = relativedelta(days=-8), 
    spread = 0.002,
    tax_model=GermanTaxModel(),
).backtest(etfs)

p_2x_sp500_ma = MAPortfolio(
    {
        "2x_sp500_eu": dict(dist=100, ma=290, ma_asset="1x_sp500_eu"),
    },
    start_value = 10000,
    tax_model=GermanTaxModel(),
).backtest(etfs)

p_hfea = MAPortfolio(
    {
        '3x_sp500_us': dict(dist=55),
        '3x_ltt_us': dict(dist=45),
    },
    start_value = 10000,
    rebalancing = relativedelta(months=3),
    rebalancing_offset = relativedelta(days=-8), 
    spread = 0.002,
    tax_model=GermanTaxModel(),
).backtest(etfs)

### Settings

start_value = 10000
monthly_rate = 100
simulation_time = relativedelta(years = 15)

### 50% Portfolio

perform_monte_carlo_simulation(
    portfolio = p_50,
    portfolio_name = '50% Portfolio',
    reference = p_cash,
    reference_name = 'cash',
    simulation_time = simulation_time,
    start_value = start_value,
    monthly_rate = monthly_rate,
    portfolio_simulations = 300,
    reference_simulations = 10,
)

### 80% Portfolio

perform_monte_carlo_simulation(
    portfolio = p_80,
    portfolio_name = '80% Portfolio',
    reference = p_cash,
    reference_name = 'cash',
    simulation_time = simulation_time,
    start_value = start_value,
    monthly_rate = monthly_rate,
    portfolio_simulations = 300,
    reference_simulations = 10,
)

### 65% (3x) Portfolio

perform_monte_carlo_simulation(
    portfolio = p_65_3x,
    portfolio_name = '65% (3x) Portfolio',
    reference = p_cash,
    reference_name = 'cash',
    simulation_time = simulation_time,
    start_value = start_value,
    monthly_rate = monthly_rate,
    portfolio_simulations = 300,
    reference_simulations = 10,
)

### HFEA

perform_monte_carlo_simulation(
    portfolio = p_hfea,
    portfolio_name = 'HFEA',
    reference = p_cash,
    reference_name = 'cash',
    simulation_time = simulation_time,
    start_value = start_value,
    monthly_rate = monthly_rate,
    portfolio_simulations = 300,
    reference_simulations = 10,
)

### 2x S&P 500 (MA) Portfolio

perform_monte_carlo_simulation(
    portfolio = p_2x_sp500_ma,
    portfolio_name = '2x S&P 500 (MA)',
    reference = p_cash,
    reference_name = 'cash',
    simulation_time = simulation_time,
    start_value = start_value,
    monthly_rate = monthly_rate,
    portfolio_simulations = 300,
    reference_simulations = 10,
)

### Wishes

perform_monte_carlo_simulation(
    portfolio = p_2x_sp500_ma,
    portfolio_name = '2x S&P 500 (MA)',
    reference = p_cash,
    reference_name = 'cash',
    simulation_time = simulation_time,
    start_value = 5800,
    monthly_rate = 250,
    portfolio_simulations = 300,
    reference_simulations = 10,
)
